Add Kuroneko_Holding_Fake_C_Book.png
This is from Oreimo Season 1 Episode 13 (depending on your episode
numbering), titled "There's No Way My Junior Can Be This Rotten". It
happens a little after 14 minutes in, in my copy, just before the the
interstitial thing before the second half of the episode.

The book does not actually exist, which I hope is ok? The readme didn't
mention anything about whether or not the book has to exist. This one
doesn't. But it's clearly about the C programming language, as the
title, "C言語", literally means "C Language".